frontend: Reorganize Resource components together
