+++ frontend/src/components/common/Resource/DeleteButton.tsx
@@ -5,9 +5,9 @@ import Tooltip from '@material-ui/core/Tooltip';
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useLocation } from 'react-router-dom';
-import { KubeObject } from '../../lib/k8s/cluster';
-import { CallbackActionOptions, clusterAction } from '../../redux/actions/actions';
-import { ConfirmDialog } from './Dialog';
+import { KubeObject } from '../../../lib/k8s/cluster';
+import { CallbackActionOptions, clusterAction } from '../../../redux/actions/actions';
+import { ConfirmDialog } from '../Dialog';
 
 interface DeleteButtonProps {
   item?: KubeObject;
