+++ frontend/src/components/common/Resource/index.tsx
@@ -1,0 +1,1 @@
+export * from './Resource';
